+++ src/models/RefreshToken.php
@@ -44,7 +44,7 @@ class RefreshToken extends BaseModel implements RefreshTokenModelInterface
     {
         $behaviors = parent::behaviors();
         $behaviors['emptyArray'] = [
-            'class' => EmptyArrayBehavior::className(),
+            'class' => EmptyArrayBehavior::class,
             'attributes' => ['scopes'],
         ];
         return $behaviors;