+++ src/services/redis/AccessTokenService.php
@@ -152,7 +152,7 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
             } catch (DatabaseException $e) {
                 // @codeCoverageIgnoreStart
                 // we have a REDIS exception, we should not discard
-                Yii::trace('Error while inserting entity', __METHOD__);
+                Yii::debug('Error while inserting entity', __METHOD__);
                 throw $e;
                 // @codeCoverageIgnoreEnd
             }
@@ -249,7 +249,7 @@ class AccessTokenService extends BaseService implements AccessTokenServiceInterf
         } catch (DatabaseException $e) {
             // @codeCoverageIgnoreStart
             // we have a REDIS exception, we should not discard
-            Yii::trace('Error while updating entity', __METHOD__);
+            Yii::debug('Error while updating entity', __METHOD__);
             throw $e;
             // @codeCoverageIgnoreEnd
         }