+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Oauth2 Yii2 Change Log
 ======================
 
+ * Enh: update minimal Yii requirement to version 2.0.13
+ * Enh: update console script deprecated error codes
+ * Enh: update deprecated method from `Yii::trace()` to `Yii::debug()`
+ * Enh: update deprecated method from `className()` to `::class`
+ * Fix: make module PHP 7.2 compliant
+ 
 1.2.1 July 11, 2017
 -------------------